% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmthemes-package.R
\docType{package}
\name{pmthemes}
\alias{pmthemes}
\alias{pmthemes-package}
\title{pmthemes}
\description{
This package contains extra themes, scales, and geoms, and
functions for and related to \pkg{ggplot2}.
}
