% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{rail_pal}
\alias{rail_pal}
\title{rail palette}
\usage{
rail_pal(palette = "aberdeen", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'rail_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
rail palette
}
\examples{
library(tidyverse)
ggplot(data = diamonds) +
geom_bar(mapping = aes(x = cut, fill = cut)) 
+ theme_wordfig() + scale_fill_rail(palette="aberdeen")

}
