% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{rail_palettes}
\alias{rail_palettes}
\title{Vintage British Rail Palettes}
\format{An object of class \code{list} of length 7.}
\usage{
rail_palettes
}
\description{
A collection of colour palettes inspired by vintage British Rail posters
}
\details{
Available palletes:

aberdeen
cornwall
harrogate
penzance
somerset
wales
yorkshirecoast
}
\examples{

# devtools::install_github("petermacp/pmthemes")

# Make an x-y plot using the aberdeen palette
library(tidyverse)
df <- data.frame(x = rnorm(100, 0, 20), 
          y = rnorm(100, 0, 20), 
          cl = sample(letters[1:8], 100, replace=TRUE))
ggplot(df, aes(x, y, colour=cl, shape=cl)) + 
  geom_point(size=4) + scale_colour_ochre() +
  theme_bw() + theme(aspect.ratio=1) 

# Make a histogram using the penzance palette
ggplot(df, aes(x, fill=cl)) + geom_histogram() +
  scale_fill_ochre(palette="penzance")

}
\keyword{datasets}
