% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pm.R
\name{theme_pm}
\alias{theme_pm}
\title{Theme for light-background presentions.}
\usage{
theme_pm(...)
}
\description{
\code{theme_pm} applies custom theme to ggplot objects.
}
\details{
This is a custom function to apply a opinionated theme to ggplot
objects destinated for scientific presentations with a light
background
}
\concept{themes}
