% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_presdark.R
\name{theme_presdark}
\alias{theme_presdark}
\title{Theme for dark-background presentions.}
\usage{
theme_presdark(...)
}
\description{
\code{theme_presdark} applies custom theme to ggplot objects.
}
\details{
This is a custom function to apply a opinionated theme to ggplot
objects destinated for scientific presentations with a dark
background
}
\seealso{
Other themes: \code{\link{theme_preslight}},
  \code{\link{theme_wordfig}}
}
