% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_preslight.R
\name{theme_preslight}
\alias{theme_preslight}
\title{Theme for light-background presentions.}
\usage{
theme_preslight(...)
}
\description{
\code{theme_preslight} applies custom theme to ggplot objects.
}
\details{
This is a custom function to apply a opinionated theme to ggplot
objects destinated for scientific presentations with a light
background
}
\seealso{
Other themes: \code{\link{theme_presdark}}
}
