% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wordfig.R
\name{theme_wordfig}
\alias{theme_wordfig}
\title{Theme for figures in word processing documents.}
\usage{
theme_wordfig(...)
}
\description{
\code{theme_wordfig} applies custom theme to ggplot objects.
}
\details{
This is a custom function to apply a opinionated theme to ggplot
objects destinated for scientific presentations with a light
background
}
\seealso{
Other themes: \code{\link{theme_presdark}},
  \code{\link{theme_preslight}}
}
\concept{themes}
