% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{rail_pal}
\alias{rail_pal}
\title{rail palette with ramped colours}
\usage{
rail_pal(palette = "aberdeen", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'rail_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
rail palette with ramped colours
}
\examples{
library(scales)
show_col(rail_pal("aberdeen")(8))

filled.contour(volcano,color.palette = rail_pal("wales"), asp=1)

}
