% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_rail}
\alias{scale_color_rail}
\alias{scale_colour_rail}
\title{Setup colour palette for ggplot2}
\usage{
scale_color_rail(..., palette = "aberdeen", discrete = TRUE,
  alpha = 1, reverse = FALSE)

scale_colour_rail(..., palette = "aberdeen", discrete = TRUE,
  alpha = 1, reverse = FALSE)
}
\arguments{
\item{...}{additional arguments to pass to scale_color_gradientn}

\item{palette}{Choose from 'rail_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{transparency}

\item{reverse}{logical, Reverse the order of the colours?}
}
\description{
Setup colour palette for ggplot2
}
\examples{
library(tidyverse)
ggplot(data = diamonds) +
geom_bar(mapping = aes(x = cut, fill = cut)) 
+ theme_wordfig() + scale_fill_rail(palette="aberdeen")

library(tidyverse)
ggplot(data = diamonds) +
geom_bar(mapping = aes(x = cut, fill = cut)) 
+ theme_wordfig() + scale_fill_rail(palette="aberdeen")
}
